requirements.txt
pytest~=7.1.2
sympy~=1.10.1
scipy~=1.8.0
pyyaml~=6.0
networkx~=2.8